LayoutTests/http/tests/xmlhttprequest/re-login-async-expected.txt

Test that default credentials aren't used when ones are provided to XHR explicitly.

Login: PASS
Default: PASS
Re-login: PASS
New default: PASS
